feat(channels): export the slash-command surface for external channel hosts
Move the pure channel slash-command surface — the command definition
list, slash/bang parsing, help text, and unknown-command messaging —
out of src/channels/commands.ts into src/channels/command-surface.ts,
and export it from the channels-public package entrypoint.

commands.ts keeps re-exporting the moved symbols and wraps the two
display-name-dependent builders with the plugin-registry resolver, so
local host behavior is byte-identical. The new module takes an
injectable ChannelDisplayNameResolver (defaulting to first-party
display names) so hosts without the plugin registry render the same
messages. Command execution stays host-injected via
ChannelSlashCommandHandlers; execution and access-control are not
exported.

Sibling of #3726: external channel hosts (Letta Cloud's Slack gateway)
can now consume the exact same command surface as `letta server
--channel` instead of maintaining their own registry.
